Resolved conflicts and adapted the branch to master's UInt256->ulong nonce /
block-number / gas migration:
- IGasPolicy / EvmInstructions.Storage: keep master's ulong arithmetic + branch's GasCostsFast alias
- VirtualMachine: adopt master's exceptionType-based out-of-gas flow, keep branch's ZK opcode-count guards and pc refactor
- StatelessExecutingWorldState: update overrides to ulong nonce / ulong CommitTree signatures
- StatelessInputGen (TransactionReader/FixtureReader): read nonce/gas/number/gasUsed as ulong

+++ src/Nethermind/Nethermind.Consensus.Ethash/HintBasedCache.cs
@@ -34,7 +34,7 @@ namespace Nethermind.Consensus.Ethash
         private readonly ILogger _logger = logManager?.GetClassLogger<HintBasedCache>() ?? throw new ArgumentNullException(nameof(logManager));
 
         [MethodImpl(MethodImplOptions.Synchronized)]
-        public void Hint(Guid guid, long start, long end)
+        public void Hint(Guid guid, ulong start, ulong end)
         {
             uint startEpoch = (uint)(start / Ethash.EpochLength);
             uint endEpoch = (uint)(end / Ethash.EpochLength);
@@ -86,7 +86,7 @@ namespace Nethermind.Consensus.Ethash
 
             if (currentMin > startEpoch || currentMax < endEpoch)
             {
-                for (long i = startEpoch; i <= endEpoch; i++)
+                for (uint i = startEpoch; i <= endEpoch; i++)
                 {
                     uint epoch = (uint)i;
                     if (epochForGuid.Add(epoch))
